Rename controller package structure
