+++ src/main/java/uk/gov/laa/ccms/caab/advice/ActiveCaseModelAdvice.java
@@ -7,20 +7,20 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import uk.gov.laa.ccms.caab.controller.AssessmentController;
-import uk.gov.laa.ccms.caab.controller.application.summary.ApplicationSectionsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.ApplicationTypeSectionController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientAddressDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientAddressDetailsSearchController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientBasicDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientContactDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientDeceasedDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientEqualOpportunitiesMonitoringDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditClientSummaryController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditGeneralDetailsSectionController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EditProceedingsAndCostsSectionController;
-import uk.gov.laa.ccms.caab.controller.application.summary.EvidenceSectionController;
-import uk.gov.laa.ccms.caab.controller.application.summary.OpponentsSectionController;
-import uk.gov.laa.ccms.caab.controller.application.summary.ProviderDetailsSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.ApplicationSectionsController;
+import uk.gov.laa.ccms.caab.controller.application.section.ApplicationTypeSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientAddressDetailsController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientAddressDetailsSearchController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientBasicDetailsController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientContactDetailsController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientDeceasedDetailsController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientEqualOpportunitiesMonitoringDetailsController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditClientSummaryController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditGeneralDetailsSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.EditProceedingsAndCostsSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.EvidenceSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.OpponentsSectionController;
+import uk.gov.laa.ccms.caab.controller.application.section.ProviderDetailsSectionController;
 
 
 /**
